Move Hebrew symbols to hebrew.toml
I think the based-on convention was a mistake. It defines things in
multiple places, making it harder to make changes. It complicates
combining config, like the ['(?!(?:LATIN|GREEK) ).*', 'CIRCLED \g<0>']
regex to only add CIRCLED characters for math symbols, not Latin or
Greek letters. And even though the math things based on Greek, Hebrew,
and Latin are actually math things, I don't think it's difficult to
remember that a symbol from Greek is in greek.toml not math.toml.

I picked "$" for the Hebrew symbols because "$" is a symbol version of
"S", sorta. I'm not sure that was a good pick, but it's easy to change
later.