Start on Runic
I only see two scripts starting with "R", so I picked a short prefix for
now. If/when other scripts are added later it might make sense to change
this prefix to something longer, to give those scripts priority for
shorter prefixes.

https://www.unicode.org/versions/Unicode16.0.0/core-spec/chapter-8/#G26624
says that the punctuation is based on graphical form more than meaning,
so I picked mnemonics that way too.