Add vocalic consonants
The Harvard-Kyoto transliterations didn't work because of a conflict
between the independent forms starting with "l" and LA followed by a
vowel sign. Instead, I tried to follow the lowercase/uppercase pattern
that Harvard-Kyoto uses elsewhere, along with Velthuis's "." before the
independent forms to prevent conflicts with consonants. It doesn't seem
great, but I'm not sure what would be better.