Split known sequences by extended grapheme clusters
The only change this has on the final map is adding LATIN CAPITAL LETTER
J + COMBINING ACUTE ACCENT and the lowercase of that, which I think is
an improvment if those are in exemplar data.

It also makes the coverage reports easier to read, since it's not that
interesting that "ij" and the like aren't covered by mnemonics.