Make expressions more explicit, and recursive
Previously, everything that returned a map computed an implicit
cartesian product, the union was implicit, and expressions couldn't
really be parsed as separate things.

Now, the product and union operations are explicit, and any expression
can be parsed and evaluated to return a map.

Aside from limited recursion (which was effectively already possible by
giving each term a name in the top-level expressions table), this commit
should not change what's possible with expressions. However, I think it
will make it easier to extend expressions in the future.

Specifically, maps (used with product) and combining.append (used with
combine) are pretty similar things now with the same structure, but they
can't be used interchangeably. When trying to add Hangul, I wanted to
use name_maps to build maps of jamo, and then use the combining.append
logic with those maps, but that's not possible because name_maps
produces a map not a combining config. With this change to how
expressions work in general, I think it will be easier to merge
combining.append with maps, and make "combine" take a map to append as a
generic expression.