Make assertions match the order they would happen
When I started on m17n-test, I was thinking about how often text is in
the pre-edit before it's committed. However, with a single sequence of
input keys, any text that's committed would generally be before any text
that's still in the pre-edit at the end of the input sequence.